Remove `dummy_order` and `dummy_line_item` helpers
These helpers are dangerous: they provide a false sense of assurance
by making you think that the order and the line item they return can
be used to infer the correct total value of future subscription orders.

In reality, order calculation in Solidus is an extremely complex process
that may take a ton of different parameters into account, and each store
is better off calculating the subscription total with their custom logic
rather than this extension trying to provide a solution that works for
everyone.

In the future, we may provide a way to compute a subscription's total,
but for the time being it's better to remove the helpers altogether.